use seperate test file, move create_multi to test file
