+++ src/aggregated_range_proof/messages.rs
@@ -71,46 +71,6 @@ pub struct Proof {
 }
 
 impl Proof {
-    pub fn create_one<R: Rng>(v: u64, n: usize, rng: &mut R) -> Proof {
-        Self::create_multi(vec![v], n, rng)
-    }
-
-    pub fn create_multi<R: Rng>(values: Vec<u64>, n: usize, rng: &mut R) -> Proof {
-        use generators::{PedersenGenerators,Generators};
-
-        let m = values.len();
-        let generators = Generators::new(PedersenGenerators::default(), n, m);
-
-        let parties: Vec<_> = values
-            .iter()
-            .map(|&v| {
-                let v_blinding = Scalar::random(rng);
-                Party::new(v, v_blinding, n, &generators)
-            })
-            .collect();
-
-        let dealer = Dealer::new(n, m).unwrap();
-
-        let (parties, value_commitments): (Vec<_>, Vec<_>) = parties
-            .iter()
-            .enumerate()
-            .map(|(j, p)| p.assign_position(j, rng))
-            .unzip();
-
-        let (dealer, value_challenge) = dealer.receive_value_commitments(&value_commitments);
-
-        let (parties, poly_commitments): (Vec<_>, Vec<_>) = parties
-            .iter()
-            .map(|p| p.apply_challenge(&value_challenge, rng))
-            .unzip();
-
-        let (dealer, poly_challenge) = dealer.receive_poly_commitments(&poly_commitments);
-
-        let proof_shares: Vec<ProofShare> = parties.iter().map(|p| p.apply_challenge(&poly_challenge)).collect();
-
-        dealer.receive_shares(&proof_shares, &generators.all(), value_challenge.y)
-    }
-
     pub fn verify<R: Rng>(&self, rng: &mut R) -> Result<(), ()> {
         use generators::{PedersenGenerators,Generators};
 
@@ -232,50 +192,4 @@ fn delta(n: usize, m: usize, y: &Scalar, z: &Scalar) -> Scalar {
     let zz = z * z;
 
     (z - zz) * sum_of_powers_of_y - z * zz * sum_of_powers_of_2 * sum_of_powers_of_z
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use rand::OsRng;
-
-    fn test_u32(m: usize) {
-        let mut rng = OsRng::new().unwrap();
-        let v: Vec<u64> = iter::repeat(())
-            .map(|()| rng.next_u32() as u64).take(m).collect();
-        let rp = Proof::create_multi(v, 32, &mut rng);
-        assert!(rp.verify(&mut rng).is_ok());
-    }
-
-    fn test_u64(m: usize) {
-        let mut rng = OsRng::new().unwrap();
-        let v: Vec<u64> = iter::repeat(())
-            .map(|()| rng.next_u64()).take(m).collect();
-        let rp = Proof::create_multi(v, 64, &mut rng);
-        assert!(rp.verify(&mut rng).is_ok());
-    }
-
-    #[test]
-    fn one_value() {
-        test_u32(1);
-        test_u64(1);
-    }
-
-    #[test]
-    fn two_values() {
-        test_u32(2);
-        test_u64(2);
-    }
-
-    #[test]
-    fn four_values() {
-        test_u32(4);
-        test_u64(4);
-    }
-
-    #[test]
-    fn eight_values() {
-        test_u32(8);
-        test_u64(8);
-    }
 }
