+++ src/aggregated_range_proof/mod.rs
@@ -3,5 +3,5 @@
 mod party;
 mod dealer;
 mod messages;
-
+mod test;
